06/requirements.txt
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn>=0.24.0
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.3.0
pydantic>=2.5.0
psutil>=5.9.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
numpy>=1.24.0
httpx>=0.25.0